test(LottoTest) : 로또 번호의 개수가 6개 보다 적거나 많다면 예외가 발생한다
